Changes to the Puppet module
This commit introduces a few changes to the Puppet module. First, the
"noop" and "apply" tasks were combined into one task. To use noop, add
it as an option to apply. Next, the masterless puppet workflow was
simplified by removing the base and roles parts and instead just
focusing on a single server manifest.